Moves XML parser part to separate directory
